requirements.txt
Flask>=2.2.3
python-dotenv>=1.0.0
openai>=1.3.0
duckduckgo-search>=3.9.3
Werkzeug>=2.2.3
gunicorn>=20.1.0 